Sort the data in several chunks
Since the data is larger than our memory, we need to sort the data
seperately.

The sort program will split the input data in every 1024 integers, and
save it in the chuck directory in ascending order.

The next step will be to merge all the chunk files into single
output.txt