Import another model of processing
I use another model to rewrite the whole structure.

In previous version, I try to seperate so called "get line" thread and
"output" thread into 2 different subroutine, but this is hard to
maintain the output order.

In this model, I try not to seperate input and output thread, and the
only concurrent part is parsing lines of data.