Fix error of duplicated output
In the scenario when multiple threads are doing the tasks, when some
threads tend to execute faster, then it may be possible to be scheduled
to do another task.

I add one more check on the scheduler to ensure the thread to be
scheduled does not have some task in pending.