Use bool variable to protect the buffer
It can be dangerous when the scheduler trying to pass a new line of data
to a working thread, and may cause the distortion of the output.

I add a bool variable into the worker_args to indicate that the thread
is working, and avoid the scheduler from pass a task on such thread.

Note: It can be some problem while using a single int variable to manage
the state of a thread, I think maybe it is occur because the assignment
cost of an integer is more larger than a bool variable.